用户登录注册、token认证、swagger接口文档

+++ disk-file-service/src/main/java/com/bumpkin/disk/file/sevice/impl/FileServiceImpl.java
@@ -1,15 +1,23 @@
 package com.bumpkin.disk.file.sevice.impl;
 
+import cn.hutool.core.io.FileUtil;
+import cn.hutool.core.util.IdUtil;
+import cn.hutool.core.util.StrUtil;
+import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
+import com.baomidou.mybatisplus.core.incrementer.IKeyGenerator;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import com.bumpkin.disk.file.dao.FileMapper;
-import com.bumpkin.disk.file.entity.File;
+import com.bumpkin.disk.file.entity.DiskFile;
 import com.bumpkin.disk.file.sevice.FileService;
-import com.bumpkin.disk.file.util.FileUtil;
+import com.bumpkin.disk.file.util.MD5Util;
+import com.bumpkin.disk.file.util.MyFileUtil;
 import com.bumpkin.disk.file.util.StringUtil;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.io.FileUtils;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
 /**
@@ -18,16 +26,20 @@ import org.springframework.web.multipart.MultipartFile;
  */
 @Slf4j
 @Service
-public class FileServiceImpl extends ServiceImpl<FileMapper, File> implements FileService {
+public class FileServiceImpl extends ServiceImpl<FileMapper, DiskFile> implements FileService {
 
     @Value("${fileRootPath}")
     public String fileRootPath;//static
+
+    @Autowired
+    public IKeyGenerator iKeyGenerator;
 
 //    @Value("${fileRootPath}")
 //    public void setFileRootPath(String fileRootPath) {
 //        FileServiceImpl.fileRootPath = fileRootPath;
 //    }
 
+    @Transactional
     @Override
     public Boolean upload(MultipartFile file, String userName, String path) {
         // 服务器上传的文件所在路径
@@ -40,13 +52,45 @@ public class FileServiceImpl extends ServiceImpl<FileMapper, File> implements Fi
         }
         // 获取上传文件的原名 例464e7a80_710229096@qq.com.zip
         String saveFileName = file.getOriginalFilename();
-        // 上传文件到-磁盘
-        try {
-            assert saveFileName != null;
-            FileUtils.copyInputStreamToFile(file.getInputStream(), new java.io.File(saveFilePath, saveFileName));
-        } catch (Exception e) {
-            log.error("Exception:", e);
-            return false;
+        assert saveFileName != null;
+        String md5ToStr = MD5Util.getFileMD5ToString(new java.io.File(saveFilePath, saveFileName));
+        if (!checkMd5Exist(md5ToStr)) {
+            // 上传文件到-磁盘
+            try {
+                FileUtils.copyInputStreamToFile(file.getInputStream(), new java.io.File(saveFilePath, saveFileName));
+                DiskFile newFile = new DiskFile();
+                String fileUuid = IdUtil.simpleUUID();
+                newFile.setId(fileUuid);
+                newFile.setFileId(fileUuid);
+                newFile.setFileLocalLocation(saveFilePath);
+                newFile.setFileSize((int) FileUtil.size(filePathDir));
+                newFile.setFileMd5(md5ToStr);
+                newFile.setFileType(StrUtil.subBefore(saveFileName, ".", false));
+                newFile.setOriginalName(saveFileName);
+                this.baseMapper.insert(newFile);
+//                //虚拟地址
+//                VirtualAddress virtualAddress = new VirtualAddress();
+//                String virtualAddressUuid = IdUtil.simpleUUID();
+//                virtualAddress.setId(virtualAddressUuid);
+//                virtualAddress.setUuid(virtualAddressUuid);
+//                virtualAddress.setFileId(fileUuid);
+//                virtualAddress.setUserId();
+//                virtualAddress.setFileName(saveFileName);
+//                virtualAddress.setAddrType();
+//                virtualAddress.setFileMd5(md5ToStr);
+//                virtualAddress.setFileSize((int) FileUtil.size(filePathDir));
+//                virtualAddress.setIsDir(0);
+//                virtualAddress.setParentPath();
+//                BaseEntity newEntity = EntityUtil.getNewEntity();
+//                virtualAddress.setCreateTime(newEntity.getCreateTime());
+//                virtualAddress.setUpdateTime(newEntity.getUpdateTime());
+            } catch (Exception e) {
+                log.error("Exception:", e);
+                return false;
+            }
+        } else {
+            //下载时如果有相同的md5值的文件则通过getFileByMd5()方法获取文件
+            //todo 上传时如果有相同的md5值的文件就只创建一条虚拟地址记录
         }
         return true;
     }
@@ -83,7 +127,7 @@ public class FileServiceImpl extends ServiceImpl<FileMapper, File> implements Fi
             oldNameWithPath = StringUtil.stringSlashToOne(fileRootPath + userName + "/" + path + "/" + oldName);
             newNameWithPath = StringUtil.stringSlashToOne(fileRootPath + userName + "/" + path + "/" + newName);
         }
-        return FileUtil.renameFile(oldNameWithPath, newNameWithPath);
+        return MyFileUtil.renameFile(oldNameWithPath, newNameWithPath);
     }
 
     @Override
@@ -93,7 +137,19 @@ public class FileServiceImpl extends ServiceImpl<FileMapper, File> implements Fi
     }
 
     @Override
+    public Boolean userFileDirMove(String fileName, String oldPath, String newPath, String userName) {
+        return null;
+    }
+
+    @Override
     public String getFileRootPath() {
         return fileRootPath;
+    }
+
+    private Boolean checkMd5Exist(String md5ToStr) {
+        QueryWrapper<DiskFile> wrapper = new QueryWrapper<>();
+        wrapper.eq("file_md5", md5ToStr);
+        DiskFile selectFile = this.baseMapper.selectOne(wrapper);
+        return selectFile != null;
     }
 }
